chore: 콘솔에서 사용되는 파일들 하나의 패키지로 이동
